use arch-agnostic `Int` instead of `Int64` (#54)

+++ src/PaddedViews.jl
@@ -267,7 +267,7 @@ end
 function _extended_axes(A, inds, dims::Tuple)
     map((r1, r2, d)->d ? r2 : r1, axes(A), inds, ntuple(i->i in collect(dims), ndims(A)))
 end
-function _extended_axes(A, inds, dims::UnitRange{Int64})
+function _extended_axes(A, inds, dims::UnitRange{Int})
     map((r1, r2, d)->d ? r2 : r1, axes(A), inds, ntuple(i->i in collect(dims), ndims(A)))
 end
